--- business-registry-dashboard/app/utils/affiliations/getAffiliationInvitationStatus.ts
@@ -1,8 +1,0 @@
-/* returns the affiliation invitation status */
-export const getAffiliationInvitationStatus = (affiliationInviteInfos: AffiliationInviteInfo[]): string => {
-  if (affiliationInviteInfos[0]?.status) {
-    return affiliationInviteInfos[0].status
-  } else {
-    return 'N/A'
-  }
-}
