+++ business-registry-dashboard/app/utils/form/handleFormInputVariant.ts
