--- business-registry-dashboard/app/utils/index.ts
@@ -1,13 +1,0 @@
-// business utils
-export { buildBusinessObject } from './business/buildBusinessObject'
-export { determineDisplayName } from './business/determineDisplayName'
-
-// name request utils
-export { buildNameRequestObject } from './name-request/buildNameRequestObject'
-export { isRejectedName } from './name-request/isRejectedName'
-export { isApprovedName } from './name-request/isApprovedName'
-
-// affiliation utils
-export { getAffiliationInvitationStatus } from './affiliations/getAffiliationInvitationStatus.js'
-export { isCurrentOrganization } from './affiliations/isCurrentOrganization.js'
-export { getRequestForAuthorizationStatusText } from './affiliations/getRequestForAuthorizationStatusText'
