--- business-registry-dashboard/app/utils/name-request/isApprovedName.ts
@@ -1,4 +1,0 @@
-/* returns true if given name is approved */
-export const isApprovedName = (name: Names): boolean => {
-  return (name.state === NrState.APPROVED)
-}
