--- business-registry-dashboard/app/utils/name-request/isRejectedName.ts
@@ -1,4 +1,0 @@
-/* returns true if the given name request is rejected */
-export const isRejectedName = (name: Names): boolean => {
-  return (name.state === NrState.REJECTED)
-}
